+++ FMS-SpringBoot-Backend/src/main/java/com/org/dao/UserDao.java
@@ -1,9 +1,8 @@
 package com.org.dao;
-import java.math.BigInteger;
 
 import org.springframework.data.repository.CrudRepository;
 import com.org.model.Users;;
 
-public interface UserDao extends CrudRepository<Users, BigInteger>{
+public interface UserDao extends CrudRepository<Users, Integer>{
 
 }
