+++ FMS-SpringBoot-Backend/src/main/java/com/org/service/UserService.java
@@ -12,9 +12,9 @@ public interface UserService {
 
 	public Users updateUser(Users newUser);
 
-	public String deleteUser(BigInteger UserId);
+	public String deleteUser(int UserId);
 
 	public Iterable<Users> displayAllUser();
 
-	public ResponseEntity<?> findUserById(BigInteger userId);
+	public ResponseEntity<?> findUserById(int userId);
 }
